Extract localisation helper method to separate file
